feat(agops)!: ec command
