+++ packages/agoric-cli/src/bin-agops.js
@@ -10,6 +10,7 @@ import anylogger from 'anylogger';
 import { Command } from 'commander';
 import path from 'path';
 import { makeOracleCommand } from './commands/oracle.js';
+import { makeEconomicCommiteeCommand } from './commands/ec.js';
 import { makePsmCommand } from './commands/psm.js';
 import { makeVaultsCommand } from './commands/vaults.js';
 import { makePerfCommand } from './commands/perf.js';
@@ -21,8 +22,9 @@ const program = new Command();
 program.name(progname).version('unversioned');
 
 program.addCommand(await makeOracleCommand(logger));
-program.addCommand(await makePsmCommand(logger));
+program.addCommand(await makeEconomicCommiteeCommand(logger));
 program.addCommand(await makePerfCommand(logger));
+program.addCommand(await makePsmCommand(logger));
 program.addCommand(await makeVaultsCommand(logger));
 
 await program.parseAsync(process.argv);
