+++ packages/agoric-cli/src/commands/psm.js
@@ -146,7 +146,7 @@ export const makePsmCommand = async logger => {
     .option('--wantMinted [DOLLARS]', 'amount of anchor tokens to give', Number)
     .option('--giveMinted [DOLLARS]', 'amount of minted tokens to give', Number)
     .option('--feePct [%]', 'Gas fee percentage', Number)
-    .option('--offerId [number]', 'Offer id', Number, Date.now())
+    .option('--offerId [string]', 'Offer id', String, `swap-${Date.now()}`)
     .action(async function (opts) {
       console.warn('running with options', opts);
       const instance = await lookupPsmInstance(opts.pair);
@@ -156,65 +156,14 @@ export const makePsmCommand = async logger => {
     });
 
   psm
-    .command('committee')
-    .description('join the economic committee')
-    .option('--offerId [number]', 'Offer id', Number, Date.now())
-    .action(async function (opts) {
-      const { economicCommittee } = agoricNames.instance;
-      assert(economicCommittee, 'missing economicCommittee');
-
-      /** @type {import('../lib/psm.js').OfferSpec} */
-      const offer = {
-        id: Number(opts.offerId),
-        invitationSpec: {
-          source: 'purse',
-          // @ts-expect-error xxx RpcRemote
-          instance: economicCommittee,
-          description: 'Voter0', // XXX it may not always be
-        },
-        proposal: {},
-      };
-
-      outputAction({
-        method: 'executeOffer',
-        offer,
-      });
-
-      console.warn('Now execute the prepared offer');
-    });
-
-  psm
-    .command('charter')
-    .description('prepare an offer to accept the charter invitation')
-    .option('--offerId [number]', 'Offer id', Number, Date.now())
-    .action(async function (opts) {
-      const { econCommitteeCharter } = agoricNames.instance;
-      assert(econCommitteeCharter, 'missing econCommitteeCharter');
-
-      /** @type {import('../lib/psm.js').OfferSpec} */
-      const offer = {
-        id: Number(opts.offerId),
-        invitationSpec: {
-          source: 'purse',
-          // @ts-expect-error xxx RpcRemote
-          instance: econCommitteeCharter,
-          description: 'charter member invitation',
-        },
-        proposal: {},
-      };
-
-      outputAction({
-        method: 'executeOffer',
-        offer,
-      });
-
-      console.warn('Now execute the prepared offer');
-    });
-
-  psm
     .command('proposePauseOffers')
     .description('propose a vote')
-    .option('--offerId [number]', 'Offer id', Number, Date.now())
+    .option(
+      '--offerId [string]',
+      'Offer id',
+      String,
+      `proposePauseOffers-${Date.now()}`,
+    )
     .requiredOption(
       '--pair [Minted.Anchor]',
       'token pair (Minted.Anchor)',
@@ -222,9 +171,8 @@ export const makePsmCommand = async logger => {
       ['IST', 'AUSD'],
     )
     .requiredOption(
-      '--psmCharterAcceptOfferId [number]',
+      '--charterAcceptOfferId [string]',
       'offer that had continuing invitation result',
-      Number,
     )
     .requiredOption(
       '--substring [string]',
@@ -243,10 +191,10 @@ export const makePsmCommand = async logger => {
 
       /** @type {import('../lib/psm.js').OfferSpec} */
       const offer = {
-        id: Number(opts.offerId),
+        id: opts.offerId,
         invitationSpec: {
           source: 'continuing',
-          previousOffer: opts.psmCharterAcceptOfferId,
+          previousOffer: opts.charterAcceptOfferId,
           invitationMakerName: 'VoteOnPauseOffers',
           // ( instance, strings list, timer deadline seconds )
           invitationArgs: harden([
@@ -270,10 +218,10 @@ export const makePsmCommand = async logger => {
     .command('proposeChangeMintLimit')
     .description('propose to change the MintLimit parameter')
     .option(
-      '--offerId [number]',
+      '--offerId [string]',
       'id of this offer (optional)',
-      Number,
-      Date.now(),
+      String,
+      `proposeChangeMintLimit-${Date.now()}`,
     )
     .requiredOption(
       '--pair [Minted.Anchor]',
@@ -282,9 +230,8 @@ export const makePsmCommand = async logger => {
       ['IST', 'AUSD'],
     )
     .requiredOption(
-      '--previousOfferId [number]',
+      '--previousOfferId [string]',
       'offer using psm charter invitation',
-      Number,
     )
     .requiredOption('--limit [number]', 'new mint limit (in IST)', Number)
     .option(
@@ -303,7 +250,7 @@ export const makePsmCommand = async logger => {
       });
       /** @type {import('../lib/psm.js').OfferSpec} */
       const offer = {
-        id: Number(opts.offerId),
+        id: opts.offerId,
         invitationSpec: {
           source: 'continuing',
           previousOffer: opts.previousOfferId,
@@ -315,70 +262,6 @@ export const makePsmCommand = async logger => {
           params: { MintLimit: scaledAmount },
           deadline: BigInt(opts.deadline * 60 + Math.round(Date.now() / 1000)),
         },
-      };
-
-      outputAction({
-        method: 'executeOffer',
-        offer,
-      });
-
-      console.warn('Now execute the prepared offer');
-    });
-
-  psm
-    .command('vote')
-    .description('vote on a question (hard-coded for now))')
-    .option('--offerId [number]', 'Offer id', Number, Date.now())
-    .requiredOption(
-      '--econCommAcceptOfferId [number]',
-      'offer that had continuing invitation result',
-      Number,
-    )
-    .requiredOption(
-      '--pair [Minted.Anchor]',
-      'token pair (Minted.Anchor)',
-      s => s.split('.'),
-      ['IST', 'AUSD'],
-    )
-    .requiredOption(
-      '--forPosition [number]',
-      'index of one position to vote for (within the question description.positions); ',
-      Number,
-    )
-    .action(async function (opts) {
-      const questionHandleCapDataStr = await vstorage.readLatest(
-        'published.committees.Economic_Committee.latestQuestion',
-      );
-      const questionDescriptions = storageHelper.unserializeTxt(
-        questionHandleCapDataStr,
-        fromBoard,
-      );
-
-      assert(questionDescriptions, 'missing questionDescriptions');
-      assert(
-        questionDescriptions.length === 1,
-        'multiple questions not supported',
-      );
-
-      const questionDesc = questionDescriptions[0];
-      // TODO support multiple position arguments
-      const chosenPositions = [questionDesc.positions[opts.forPosition]];
-      assert(chosenPositions, `undefined position index ${opts.forPosition}`);
-
-      /** @type {import('../lib/psm.js').OfferSpec} */
-      const offer = {
-        id: Number(opts.offerId),
-        invitationSpec: {
-          source: 'continuing',
-          previousOffer: opts.econCommAcceptOfferId,
-          invitationMakerName: 'makeVoteInvitation',
-          // (positionList, questionHandle)
-          invitationArgs: harden([
-            chosenPositions,
-            questionDesc.questionHandle,
-          ]),
-        },
-        proposal: {},
       };
 
       outputAction({
